search: make search tree a graph and re-use already created nodes
If a successor node contains the same canonical words as another node,
do not recreate the node but instead re-use the previously created node.
Globally store all created nodes in a map canonical words -> node. This
avoids re-creating sub-trees over and over again.

This fundamentally changes the search tree to a (possibly cyclic) graph.
In particular, a node may have multiple parents.

This also breaks several things that still need to be fixed:
* Incoming actions need to be specific for the parent where the
  transition originated.
* The monotonic domination check is broken and currently completely
  disabled.
* Creating a controller is broken, e.g., if the search graph contains
  self loops. In this case, no label is being set.
* Tree iteration does not work anymore.
* Printing a tree recursively does not work anymore.

Some of those issues need to be fixed, others may not be necessary
anymore.